+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Added:
 - CTCP requests through commands and user context
 - Title bar button and keyboard shortcut to mark a buffer as read (will update the read marker as well, if the `read-marker` capability is available)
 - Mark as Read settings to control when buffers are automatically marked as read
+- `/hop` command. `/hop` parts the current channel and joins a new one
 
 Changed:
 - Clicking to insert a username will now use same suffixes specified for autocomplete